Separate functional tests from loading tests.
